Reorganize plasma wakefields
